Particle Control Panel
Control panel composed of 177 primitive objects (fashioned into buttons & displays) that communicates with the Particle Emitter to create, edit, and view particle effects in real time, saving the hassle of constant script editing and item re-spawning. There are parts that could be condensed, and it's not idiot-proof, but all the functions work. To use custom textures, click the "Define" button and paste the texture's UUID into chat. Same thing to define an emitter target. For the display, a green indicator means positive value, red means negative. Change INC up or down to describe how much the values are altered per click.