Color-changing torch script
I used to make decorative torches and fountains with it, but this script can be placed in pretty much anything to make it appear to have caught fire. The default particle values produce a pleasing flame effect without resorting to textures. Can be turned on/off, burn at a set color, and can be programmed via chat with a list of your chosen colors at your chosen ms interval. Includes functionality to allow access by owner only, or to the public. Instructions are found in 'torch-inst' and the default color list to load can be found in 'torch-colors'.