Christmas present
An item I made as a way to give gifts to people ahead of the holidays. The recipient cannot see the contents and cannot receive the gift until the declared date. Only the intended recipient can open this gift! Instructions are in 'xmas-inst', if you don't set the date it just opens as long as the date is after 12/25/2004 :P